+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-input/example/HideRequiredDateInputField.jsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-
 import DateInputField from 'terra-date-input/lib/DateInputField';
 
 const Example = () => {
@@ -9,13 +8,13 @@ const Example = () => {
   return (
     <React.Fragment>
       <DateInputField
-        legend="Legend text"
+        legend="Select Event Date"
         name="date-input-value"
         value={value}
         onChange={(event, dateString) => setValue(dateString)}
         required
         hideRequired
-        error="Error message"
+        error="Please enter a valid event date."
         help="Help message"
         isInvalid={isInvalid}
       />